SY-4221: Move text dimensions out of @synnaxlabs/x into pluto text/aether
