+++ pluto/src/lineplot/Axis.tsx
@@ -9,7 +9,7 @@
 
 import "@/lineplot/Axis.css";
 
-import { type bounds, direction, text } from "@synnaxlabs/x";
+import { type bounds, direction, type text } from "@synnaxlabs/x";
 import {
   type FC,
   type PropsWithChildren,
@@ -32,6 +32,7 @@ import {
 import { useGridEntry } from "@/lineplot/LinePlot";
 import { useMemoDeepEqual } from "@/memo";
 import { Text } from "@/text";
+import { text as aetherText } from "@/text/aether";
 import { Theming } from "@/theming";
 
 export interface AxisProps
@@ -119,12 +120,12 @@ export const axisFactory = (dir: direction.Direction): FC<AxisProps> => {
     useEffect(() => {
       if (dir === "y") {
         if (label == null) return;
-        const dims = text.dimensions(label, font);
+        const dims = aetherText.dimensions(label, font);
         let labelSize = dims[direction.dimension(direction.construct(labelDirection))];
         if (labelSize > 0) labelSize += 9;
         setState((state) => ({ ...state, labelSize }));
       } else {
-        const dims = text.dimensions(label, font);
+        const dims = aetherText.dimensions(label, font);
         let labelSize = dims.height * 1.3;
         if (labelSize > 0) labelSize += 12;
         const prevSize = prevLabelSize.current;
