+++ pluto/src/text/aether/dimensions.ts
@@ -7,8 +7,7 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-import { runtime } from "@/runtime";
-import { type dimensions as base } from "@/spatial/dimensions";
+import { type dimensions as base, runtime } from "@synnaxlabs/x";
 
 let canvas: HTMLCanvasElement | null = null;
