+++ pluto/src/text/aether/external.ts
@@ -7,5 +7,5 @@
 // License, use of this software will be governed by the Apache License, Version 2.0,
 // included in the file licenses/APL.txt.
 
-export * from "@/text/types";
-export * from "@/text/types.gen";
+export * from "@/text/aether/atlas";
+export * from "@/text/aether/dimensions";
