+++ pluto/src/vis/render/context.ts
@@ -22,11 +22,11 @@ import { type aether } from "@/aether/aether";
 import { alamos } from "@/alamos/aether";
 import { CSS } from "@/css";
 import { status } from "@/status/aether";
+import { text } from "@/text/aether";
 import { SugaredOffscreenCanvasRenderingContext2D } from "@/vis/draw2d/canvas";
 import { clear } from "@/vis/render/clear";
 import { Loop } from "@/vis/render/loop";
 import { applyOverScan } from "@/vis/render/util";
-import { text } from "@/vis/text";
 
 export type Canvas2DVariant = "upper2d" | "lower2d";
 export type CanvasGLVariant = "gl";
